refactor(carbon): move NSVB scalar test-oracle out of src into tests (#127.1)
The production NSVB path is fully vectorized (polars). A parallel *scalar*
reimplementation of the same math shipped in src/ purely as a test oracle
(equations.py Models 1-5 / predict_tree_biomass / harmonize_components /
_eval_component / Coefficients / TreeBiomassResult; coefficients.py
lookup_coefficients / _row_to_dict / ecosubcd_to_division). Move it to
tests/nsvb_oracle.py so the shipped library no longer carries a second
implementation, wired via `pythonpath = ["tests"]`.

The oracle is kept (not deleted): test_nsvb_vectorized.py cross-checks the
vectorized pipeline tree-for-tree against this independent implementation, and
both are anchored to the hand-computed GTR-WO-104 worked-example values in
test_nsvb_equations.py — a cross-check that has caught real bugs (the #123
half-implemented Model 2). Removing it would trade away that independent oracle.

Scope note: the carbon_fractions.py dict loaders were left in src. They looked
test-only but are production dependencies — `_compute_default_{live,dead}_carbon_fraction`
(used by live_tree/standing_dead) derive their means from them — so only the
equations + coefficients scalar reference moved. Verified before removing, per
the "check the coupling first" guidance.

No production code path changed: shared constants (`_K_*`, `_CULL_DENS_PROP`),
`load_nsvb_coefficients`/`CoefficientTables`, the `build_*`/`get_vectorized_lookup_tables`
loaders, `ecosubcd_to_division_expr`, and the `_df` fraction loaders all stay.
Carbon point estimates are byte-identical on GA/NV (ag+dead). 924 unit + 12
property/validation tests pass; ruff + mypy clean.
