+++ src/pyfia/carbon/nsvb/coefficients.py
@@ -322,71 +322,6 @@ def build_stdorg_lookup(table_spcd: pl.DataFrame) -> pl.DataFrame:
     ).select(["SPCD", "STDORGCD", *_VECTORIZED_COEF_COLS])
 
 
-def ecosubcd_to_division(ecosubcd: str | None) -> str | None:
-    """Extract the Bailey DIVISION code from a ``PLOTGEOM.ECOSUBCD`` value.
-
-    ``ECOSUBCD`` is a 5-7 character Bailey ecoprovince subsection code
-    (e.g., ``"231Ae"`` for the Southeastern Mixed Forest section 231A,
-    subsection 231Ae, or ``"M231Aa"`` for the Ouachita Mixed Forest
-    mountain variant). The Bailey hierarchy is:
-
-    ::
-
-        Domain → Division → Province → Section → Subsection
-        200    → 230      → 231      → 231A    → 231Ae
-
-    This function walks one level up the hierarchy from Subsection to
-    Division by:
-
-    - extracting the 3-digit Province code (first 3 chars after any ``M``)
-    - replacing its last digit with ``"0"`` to obtain the Division
-    - preserving the ``"M"`` prefix for mountain Divisions
-
-    Examples
-    --------
-    >>> ecosubcd_to_division("231Ae")
-    '230'
-    >>> ecosubcd_to_division("232Bh")
-    '230'
-    >>> ecosubcd_to_division("M231Aa")
-    'M230'
-    >>> ecosubcd_to_division("220Eb")
-    '220'
-    >>> ecosubcd_to_division(None) is None
-    True
-    >>> ecosubcd_to_division("") is None
-    True
-    >>> ecosubcd_to_division("XYZ") is None
-    True
-
-    Parameters
-    ----------
-    ecosubcd : str or None
-        The ECOSUBCD value from ``PLOTGEOM.ECOSUBCD`` or ``PLOT.ECOSUBCD``
-        (the column exists on both tables — pyfia's DataMart CSV downloads
-        pull it reliably from ``PLOTGEOM``).
-
-    Returns
-    -------
-    str or None
-        The Bailey DIVISION code matching the ``DIVISION`` column in the
-        NSVB coefficient tables (e.g., ``"230"``, ``"M230"``, ``"240"``),
-        or ``None`` if the input is null, empty, or malformed.
-    """
-    if ecosubcd is None:
-        return None
-    s = ecosubcd.strip().upper()
-    if not s:
-        return None
-    m_prefix = ""
-    if s.startswith("M"):
-        m_prefix = "M"
-        s = s[1:]
-    if len(s) < 3 or not s[:2].isdigit() or not s[2].isdigit():
-        return None
-    return m_prefix + s[:2] + "0"
-
-
 def ecosubcd_to_division_expr(col: str = "ECOSUBCD") -> pl.Expr:
     """Vectorized Polars expression equivalent of :func:`ecosubcd_to_division`.
 
@@ -531,107 +466,3 @@ def load_nsvb_coefficients() -> CoefficientTables:
                     df = df.with_columns(pl.lit(None, dtype=pl.Float64).alias(col))
         loaded[key] = df
     return CoefficientTables(**loaded)
-
-
-def _row_to_dict(row: pl.DataFrame, source: str) -> dict:
-    """Convert a single-row Polars DataFrame to a coefficient dict.
-
-    Returns the row's columns as a plain Python dict, with NaN/null values
-    coerced to 0.0 for numeric fields and the lookup ``source`` tag attached.
-    """
-    raw = row.to_dicts()[0]
-    out: dict = {"source": source}
-    for key in ("model", "a", "a1", "b", "b1", "c", "c1"):
-        val = raw.get(key)
-        if val is None:
-            out[key] = 0.0
-        else:
-            out[key] = float(val)
-    out["model"] = int(out["model"])
-    return out
-
-
-def lookup_coefficients(
-    table_spcd: pl.DataFrame,
-    table_jenkins: pl.DataFrame,
-    spcd: int,
-    jenkins_spgrpcd: int | None = None,
-    division: str | None = None,
-    stdorgcd: int | None = None,
-) -> dict:
-    """Resolve a coefficient row for one tree following NSVB lookup precedence.
-
-    The Phase 1 implementation walks precedence levels 1-4 in order. Phase 1
-    typically only hits levels 3 and 4: the SPCD-only species-level row, or
-    the Jenkins fallback for unsupported species. The DIVISION-specific levels
-    1 and 2 are wired but unused until pyFIA gains a ``PLOT.ECOSUBCD →
-    DIVISION`` mapping.
-
-    Parameters
-    ----------
-    table_spcd : pl.DataFrame
-        The ``S*a`` species-keyed table (e.g., ``volib_spcd``).
-    table_jenkins : pl.DataFrame
-        The ``S*b`` Jenkins-keyed fallback table.
-    spcd : int
-        FIA species code from ``TREE.SPCD``.
-    jenkins_spgrpcd : int, optional
-        Jenkins species group code from ``REF_SPECIES.JENKINS_SPGRPCD``.
-        Required if the species is not in the SPCD table (level 4 fallback).
-    division : str, optional
-        Bailey ecoprovince division code (e.g., ``"M240"``). Phase 1 always
-        passes ``None`` here.
-    stdorgcd : int, optional
-        Stand origin code from ``COND.STDORGCD``. Phase 1 always passes ``None``.
-
-    Returns
-    -------
-    dict
-        Coefficient dict with keys ``model, a, a1, b, b1, c, c1, source``.
-        ``source`` is one of ``"spcd_division_stdorg"``, ``"spcd_division"``,
-        ``"spcd"``, or ``"jenkins"``.
-
-    Raises
-    ------
-    KeyError
-        If neither the SPCD nor any Jenkins fallback can be resolved.
-    """
-    # TODO(PR 2): Scalar reference implementation. The vectorized
-    # LiveTreeEstimator must replace per-tree calls with a polars join on
-    # SPCD against the coefficient tables, not call this function in a loop.
-    # See `pyfia/carbon/__init__.py` "Architectural rules" rule 2.
-    df = table_spcd.filter(pl.col("SPCD") == spcd)
-
-    # Level 1: SPCD + DIVISION + STDORGCD exact match
-    if division is not None and stdorgcd is not None:
-        match = df.filter(
-            (pl.col("DIVISION") == division) & (pl.col("STDORGCD") == stdorgcd)
-        )
-        if match.height > 0:
-            return _row_to_dict(match.head(1), "spcd_division_stdorg")
-
-    # Level 2: SPCD + DIVISION (STDORGCD null)
-    if division is not None:
-        match = df.filter(
-            (pl.col("DIVISION") == division) & pl.col("STDORGCD").is_null()
-        )
-        if match.height > 0:
-            return _row_to_dict(match.head(1), "spcd_division")
-
-    # Level 3: SPCD only (DIVISION and STDORGCD both null) — the species-level row
-    match = df.filter(pl.col("DIVISION").is_null() & pl.col("STDORGCD").is_null())
-    if match.height > 0:
-        return _row_to_dict(match.head(1), "spcd")
-
-    # Level 4: Jenkins fallback. Note that S*b tables have different schemas
-    # (no DIVISION/STDORGCD columns) and are keyed on JENKINS_SPGRPCD.
-    if jenkins_spgrpcd is not None:
-        jdf = table_jenkins.filter(pl.col("JENKINS_SPGRPCD") == jenkins_spgrpcd)
-        if jdf.height > 0:
-            return _row_to_dict(jdf.head(1), "jenkins")
-
-    raise KeyError(
-        f"No NSVB coefficients for SPCD={spcd}, JENKINS_SPGRPCD={jenkins_spgrpcd}, "
-        f"DIVISION={division}, STDORGCD={stdorgcd}. Check the species coverage "
-        "in src/pyfia/carbon/nsvb/data/."
-    )
